Generate a pkl file to save filenames and classes
When ImageDataGeneratorCustom is called and a lot of images and classes are used in training (i.e. market 1501 dataset) loading of classes and filenames can be quite long. 

Hear, I save those on a pickle file for use next time deepRanking.py is called, in particular when the triplets generated have not changed.

If triplets are new, then generated 'delete_if_new_data_source.pkl' should be deleted.